refactor(playlist): extract evaluator to internal/playlist package
Move playlist_evaluator.go and playlist_evaluator_test.go from internal/server
to internal/playlist. The evaluator tests now run in the playlist package context
rather than through the HTTP stack. Update playlist_handlers.go to delegate to
the extracted evaluator via the playlist package.
